#1180 Move Menas plugin for Atum to the Spark Jobs module.
