+++ dao/src/main/scala/za/co/absa/enceladus/dao/auth/MenasCredentialsFactory.scala
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package za.co.absa.enceladus.dao.menasplugin
+package za.co.absa.enceladus.dao.auth
 
 import org.apache.spark.sql.SparkSession
