+++ dao/src/main/scala/za/co/absa/enceladus/dao/rest/AuthClient.scala
@@ -20,7 +20,7 @@ import org.springframework.http.{HttpHeaders, HttpStatus, ResponseEntity}
 import org.springframework.security.kerberos.client.KerberosRestTemplate
 import org.springframework.util.LinkedMultiValueMap
 import org.springframework.web.client.RestTemplate
-import za.co.absa.enceladus.dao.menasplugin._
+import za.co.absa.enceladus.dao.auth._
 import za.co.absa.enceladus.dao.UnauthorizedException
 
 object AuthClient {
