+++ spark-jobs/src/main/scala/za/co/absa/enceladus/conformance/ConfCmdConfig.scala
@@ -17,7 +17,7 @@ package za.co.absa.enceladus.conformance
 
 import org.apache.spark.storage.StorageLevel
 import scopt.OptionParser
-import za.co.absa.enceladus.dao.menasplugin.{InvalidMenasCredentialsFactory, MenasCredentialsFactory, MenasKerberosCredentialsFactory, MenasPlainCredentialsFactory}
+import za.co.absa.enceladus.dao.auth._
 
 import scala.util.matching.Regex
