+++ spark-jobs/src/test/scala/za/co/absa/enceladus/conformance/ConfConfigSuite.scala
@@ -18,7 +18,7 @@ package za.co.absa.enceladus.conformance
 import java.time.ZonedDateTime
 
 import org.scalatest.FunSuite
-import za.co.absa.enceladus.dao.menasplugin.{MenasKerberosCredentials, MenasPlainCredentials}
+import za.co.absa.enceladus.dao.auth.{MenasKerberosCredentials, MenasPlainCredentials}
 import za.co.absa.enceladus.model.Dataset
 import za.co.absa.enceladus.utils.testUtils.SparkTestBase
